Refactor project tests and request factories
